refactor(ui): Extract form schema
These functions and types relate to creation of the schema of the ORT run
creation form, and handling its error states.

Note that the awkward import / then re-export structure in this and two
following commits is just a temporary nuisance, to keep the call sites of
the refactored methods happy, and will vanish by the end of this refactor.
